yasdi: add also .so links in main package, to accomodate our usage
By default Yocto will put the .so links only in the -dev packages,
so runtime needs to pin to a specific version (ie. link to .so.X).
This commit adds the .so in the main package (and removes from dev
accordingly not to have a conflict, plus shuts up the QA check dev-so
that check for this kind of this) to align to our current usage
of the library.
We may revert this one day in the future in sync with change of how
the Yasdi code parts are built.

Additionally: just for distibuting simplicity bumped from r0 to r1
so it is easier to keep packages versions under control